aaaaaaaaaaa

+++ database/migrations/2025_09_04_235156_create_publications_table.php
@@ -14,7 +14,7 @@ return new class extends Migration
         Schema::create('publications', function (Blueprint $table) {
             $table->id();
              $table->string('titulo', 255);
-            $table->integer('precio')->nullable();
+            $table->decimal('precio')->nullable();
             $table->text('descripcion')->nullable();
             $table->boolean('visibilidad')->default(true);
